// src/core/pattern-message.enum.ts
export enum EPatternMessage {
  INFO_CUSTOMER = 'mp_info_customer',
  INFO_PARTNER_KEY = 'mp_info_partner_key',
  CALCULATE_LOYALTY_POINT = 'mp_calculate_loyalty_point',
  CALCULATE_TRANSACTION_POINT = 'mp_calculate_transaction_point',
  WRITE_WALLET = 'ep_write_wallet',
  SEND_EMAIL = 'ep_send_mail',
}